payment-request/PaymentRequestUpdateEvent/updateWith-state-checks-manual.https.html

<!doctype html>
<meta charset="utf8">
<link rel="help" href="https://www.w3.org/TR/payment-request/#updatewith()-method">
<title>
</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({ explicit_done: true, explicit_timeout: true });
const validMethod = Object.freeze({ supportedMethods: "basic-card" });
const validMethods = Object.freeze([validMethod]);
const validAmount = Object.freeze({ currency: "USD", value: "5.00" });
const validTotal = Object.freeze({
  label: "label",
  amount: validAmount,
});
const validShippingOption = Object.freeze({
  id: "a-shipping-option",
  label: "A shipping option",
  amount: validAmount,
  selected: true,
});
const validDetails = Object.freeze({
  total: validTotal,
  shippingOptions: [validShippingOption],
});
const validOptions = Object.freeze({
  requestShipping: true,
});

function getPaymentPromises() {
  const request = new PaymentRequest(validMethods, validDetails, validOptions);
  const eventPromise = new Promise(resolve => {
    request.addEventListener("shippingaddresschange", resolve);
  });
  const responsePromise = request.show();
  return { eventPromise, responsePromise };
}

function testRequestIsClosed({ textContent: testName }) {
  promise_test(async t => {
    const { eventPromise, responsePromise } = getPaymentPromises();
    const event = await eventPromise;
    // Set request.[[state]] to closed
    const abortPromise = event.target.abort();
    assert_throws(
      "InvalidStateError",
      () => {
        event.updateWith(validDetails);
      },
      "request.[[state]] is not interactive, must throw an InvalidStateError."
    );
    await abortPromise;
    responsePromise.catch(err => err);
  }, testName.trim());
}

function testRequestIsUpdating({ textContent: testName }) {
  promise_test(async t => {
    const { eventPromise, responsePromise } = getPaymentPromises();
    const event = await eventPromise;
    // Set request.[[updating]] to true
    event.updateWith(validDetails);
    assert_throws(
      "InvalidStateError",
      () => {
        event.updateWith(validDetails);
      },
      "request.[[updating]] to true, must throw an InvalidStateError."
    );
    await event.target.abort();
    responsePromise.catch(err => err);
  }, testName.trim());
}
</script>
<h2>updateWith() method - state machine checks</h2>
<p>
  When the payment sheet is shown, select a different shipping address once. Then pay.
</p>
<ol>
  <li id="test-0">
    <button onclick="testRequestIsClosed(this);">
      When updateWith() is called, if request.[[state]] is not "interactive", then throw an " InvalidStateError" DOMException.
    </button>
  </li>
  <li id="test-1">
    <button onclick="testRequestIsUpdating(this);">
      When updateWith() is called, If request.[[updating]] is true, then throw an "InvalidStateError" DOMException.
    </button>
  </li>
  <li>
    <button onclick="done();">Done!</button>
  </li>
</ol>
